E-commerce Checkout Cart
The following file contains the code developed for the e-commerce product check-out cart using PHP and MySql.

The Code connects to the database which contains a tablw with all the listed workout products, their price and description. These products are then presented on the page and the user is able to add the products they want and then view their cumulative pricing.

Attached are also images that will be used as product place holders and pulled from the code provided.